Separated upload button from dev panel
