+++ packages/frontend-react/src/features/developerPanel/DeveloperPanel.tsx
@@ -1,25 +1,10 @@
-import { Button, Drawer, Text } from "@mantine/core";
+import { Button, Drawer } from "@mantine/core";
 import { useDisclosure } from "@mantine/hooks";
-import { modals } from "@mantine/modals";
 import styles from "./DeveloperPanel.module.css";
-import UploadTransmission from "./UploadTransmission";
-
-/** The ID to use when opening / closing the UploadTransmission modal. */
-const uploadModalId = "uploadTransmissionModal";
+import UploadTransmissionButton from "./UploadTransmissionButton";
 
 export default function DeveloperPanel() {
   const [opened, { open, close }] = useDisclosure(false);
-
-  const closeUploadTransmission = () => {
-    modals.close(uploadModalId);
-  };
-
-  const openUploadTransmission = () =>
-    modals.open({
-      title: <Text size="lg">Upload a new Transmission to `emptystream`</Text>,
-      children: <UploadTransmission onComplete={closeUploadTransmission} />,
-      modalId: uploadModalId,
-    });
 
   return (
     <div className={styles.container}>
@@ -30,7 +15,7 @@ export default function DeveloperPanel() {
         position="right"
         overlayProps={{ backgroundOpacity: 0.5, blur: 2 }}
       >
-        <Button onClick={openUploadTransmission}>Upload Transmission</Button>
+        <UploadTransmissionButton />
       </Drawer>
 
       <Button className={styles.devPanelButton} onClick={open}>
